Changed ptv client to ignore unknown properties in generated models
