+++ pom.xml
@@ -62,7 +62,7 @@
 
     <!-- Kunta API Spec -->
 
-    <kunta-api-spec.version>0.1.108</kunta-api-spec.version>
+    <kunta-api-spec.version>0.1.110</kunta-api-spec.version>
     
     <!-- Integrations -->
