+++ src/main/java/fi/otavanopisto/kuntaapi/server/integrations/BinaryHttpClient.java
@@ -28,7 +28,7 @@ import org.apache.http.impl.client.BasicCredentialsProvider;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClients;
 
-import fi.otavanopisto.kuntaapi.server.integrations.GenericHttpClient.Response;
+import fi.otavanopisto.kuntaapi.server.integrations.AbstractHttpClient.Response;
 
 /**
  * HTTP client for downloading binaries
