+++ src/main/java/fi/otavanopisto/kuntaapi/server/integrations/GenericHttpCache.java
@@ -16,8 +16,8 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 
-import fi.otavanopisto.kuntaapi.server.integrations.GenericHttpClient.Response;
-import fi.otavanopisto.kuntaapi.server.integrations.GenericHttpClient.ResultType;
+import fi.otavanopisto.kuntaapi.server.integrations.AbstractHttpClient.Response;
+import fi.otavanopisto.kuntaapi.server.integrations.AbstractHttpClient.ResultType;
 
 /**
  * Response aware client cache for integrations
