+++ src/main/java/fi/otavanopisto/kuntaapi/server/integrations/ptv/updaters/PtvAccessTokenUpdater.java
@@ -25,8 +25,8 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import fi.otavanopisto.kuntaapi.server.discover.EntityUpdater;
 import fi.otavanopisto.kuntaapi.server.id.OrganizationId;
 import fi.otavanopisto.kuntaapi.server.integrations.GenericHttpClient;
-import fi.otavanopisto.kuntaapi.server.integrations.GenericHttpClient.Response;
-import fi.otavanopisto.kuntaapi.server.integrations.GenericHttpClient.ResultType;
+import fi.otavanopisto.kuntaapi.server.integrations.AbstractHttpClient.Response;
+import fi.otavanopisto.kuntaapi.server.integrations.AbstractHttpClient.ResultType;
 import fi.otavanopisto.kuntaapi.server.integrations.ptv.PtvConsts;
 import fi.otavanopisto.kuntaapi.server.integrations.ptv.tasks.AccessTokenTaskQueue;
 import fi.otavanopisto.kuntaapi.server.security.ExternalAccessTokenController;
