+++ src/main/java/fi/otavanopisto/kuntaapi/server/rest/CodesApiImpl.java
@@ -12,6 +12,7 @@ import javax.inject.Inject;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.Request;
 import javax.ws.rs.core.Response;
+import javax.ws.rs.core.Response.Status;
 
 import org.apache.commons.lang3.EnumUtils;
 import org.apache.commons.lang3.StringUtils;
@@ -19,12 +20,14 @@ import org.apache.commons.lang3.StringUtils;
 import fi.metatavu.kuntaapi.server.rest.CodesApi;
 import fi.metatavu.kuntaapi.server.rest.CodesApi;
 import fi.metatavu.kuntaapi.server.rest.model.Code;
+import fi.metatavu.kuntaapi.server.rest.model.Organization;
 import fi.otavanopisto.kuntaapi.server.controllers.CodeController;
 import fi.otavanopisto.kuntaapi.server.controllers.HttpCacheController;
 import fi.otavanopisto.kuntaapi.server.controllers.CodeController;
 import fi.otavanopisto.kuntaapi.server.id.OrganizationId;
 import fi.otavanopisto.kuntaapi.server.id.CodeId;
 import fi.otavanopisto.kuntaapi.server.integrations.KuntaApiConsts;
+import fi.otavanopisto.kuntaapi.server.integrations.KuntaApiIdFactory;
 import fi.otavanopisto.kuntaapi.server.integrations.CodeSortBy;
 import fi.otavanopisto.kuntaapi.server.integrations.CodeSortBy;
 import fi.otavanopisto.kuntaapi.server.integrations.SortDir;
@@ -41,7 +44,11 @@ public class CodesApiImpl extends CodesApi {
 
   private static final String INVALID_VALUE_FOR_SORT_DIR = "Invalid value for sortDir";
   private static final String INVALID_VALUE_FOR_SORT_BY = "Invalid value for sortBy";
+  private static final String NOT_FOUND = "Not Found";
   
+  @Inject
+  private KuntaApiIdFactory kuntaApiIdFactory;
+
   @Inject
   private CodeController codeController;
   
@@ -53,6 +60,26 @@ public class CodesApiImpl extends CodesApi {
   
   @Inject
   private RestResponseBuilder restResponseBuilder;
+  
+  @Override
+  public Response findCode(String codeIdParam, Request request) {
+    CodeId codeId = kuntaApiIdFactory.createCodeId(codeIdParam);
+    if (codeId == null) {
+      return createNotFound(NOT_FOUND);
+    }
+    
+    Response notModified = httpCacheController.getNotModified(request, codeId);
+    if (notModified != null) {
+      return notModified;
+    }
+    
+    Code code = codeController.findCode(codeId);
+    if (code != null) {
+      return httpCacheController.sendModified(code, code.getId());
+    }
+      
+    return createNotFound(NOT_FOUND);
+  }
   
   @Override
   public Response listCodes(List<String> types, String search, String sortByParam, String sortDirParam, Long firstResult, Long maxResultsParam, Request request) {
